Add support for three-part image
This patch adds support to snaps for three-part images.
The 4 new properies are added to the image:
parts: sais if this is a 1-part or a 3-part image
kernel: the location of the kernel image
ramdisk: the location of the initramfs image
extra-properties: generic field to pass other properties of the image

The kernel and the ramdisk fields can be either urls or files.
The assumption is that when dealing with 3-part images the kernel and
ramdisk are taken from the same source type as the main image.

JIRA: SNAPS-32

Change-Id: Ifb53d1da1085fcd6429ddc0607c905522db5e8bb
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>