Add support for extra properties in an image
This patch adds a new setting for images, extra_properties.
This field is a dict and can be used to create a 3-part image
by setting the kernel_id and the ramdisk_id of the main image.

Unit tests have been added to set some generic property, for
creation of the 3-part image, and for creating an instance
with that 3-part image

JIRA: SNAPS-32

Change-Id: Ifb53d1da1085fcd6429ddc0607c905522db5e8bb
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>