Added the ability to give the tests the ability to add in flavor metadata.
NFVI's configured for OVS/DPDK and fd.io scenarios required special
metadata for flavors in order for VM instances to be activated. This
patch allows for setting a value to be applied to all tests. With
this change, the default value is being removed that only partially
addressed this issue.

JIRA: SNAPS-38

Change-Id: Ic42944a849f710f631fe7ac6fbefb720a73f0b77
Signed-off-by: spisarski <s.pisarski@cablelabs.com>