Adapt integration tests to use custom image config
This fix uses the possibility to pass custom imformation about the
image via image_metadata variable. In the integration tests it checks
for this info and creates a 3-part image if the kernel_url and
ramdisk_url are set. The fix also allows the default url to be overriden
as well as passing extra-properties. All the metadata info comes from
the app.

The part in functest where the custom config is set for creating a 3-part
image is https://gerrit.opnfv.org/gerrit/29987

Change-Id: Ib2ad823705dbea68d672514c7c89b150c9c6dc63
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>