Add DEBUG logs to SimpleHealthCheck for VM console output.
The test in the class SimpleHealthCheck uses the console output
to determine whether or not the VM obtains a DHCP lease which
contains a very image-specific check. When a client changes
the image from what the test was designed for, the console
output can change causing the test to fail unless it is simply
timing out. This patch will output the full console log on
failure or on success when the logger is configured as DEBUG.

JIRA: SNAPS-47

Change-Id: I2041ca1aed8bd1e2fb65932591f4acbd4d3cd720
Signed-off-by: spisarski <s.pisarski@cablelabs.com>