Update patch set 2
Patch Set 2:

(1 comment)

Steven, First of all, thanks for the review! I would like to add some similar modifications for the other optional variables as well, so there's no hurry to merge the patch. As soon as that happens, I will also try to manually test the new version of the patch. Now I do not have a POD to test it anymore, but I have manually tested the current version of the patch and it worked as expected. Thanks!

Patch-set: 2
Label: Verified=0