Update patch set 1
Patch Set 1:

I am unsure whether or not 3-part images are working properly as I have had runtime issues when adding in the extra_properties with the Glance v2 client. The test that creates an instance from a cirros 3-part image is still working but the disk image is the same one I am using for standard single-part images. As the main use case for them is for ARM based clouds and I do not have one, can somebody from the ARM team check to see if they are still working as expected?

Patch-set: 1
Label: Verified=0