Update patch set 1
Patch Set 1:

I don't think we must react to every side effects during the early phase of the release. We have been very stable (from a master point of view) from the beginning of E and we shouldn't multiple incomplete reviews to be fixed later. MANIFEST.IN is useless if we switch to pbr as functest did.

Tim. Could you please give us at least one day to complete that correctly? We are doing big changes on functest about how we handle requirements. And all snaps issues are raised due to this change.

Patch-set: 1