Update patch set 2
Patch Set 2:

> BTW, the failure/stack trace occurred on a virtual Apex CO pod...

Actually the function of enabling snaps work using public endpoint is only for v3 when OS_INERFACE is set to be public manually, but this patch should not block the existing tests for snaps.

I guess public endpoint is not accessible for apex v2 according to the stack trace. Thank you Steven for telling me that. I will try to find an apex SUT with identity v2 to have a test.

Patch-set: 2
Label: Verified=0