Update patch set 1
Patch Set 1:

(1 comment)

Linda, the potential fix to your issue basically lies in line 153 of os_source_file_test.py. The rest of the changes in that file are for pep8 formatting only. The other files were changed merely to support adding a role to a user for v2 & v3.

Patch-set: 1