Update patch set 1
Patch Set 1:

As I need to kickoff the process from the jenkins slave and execute on a separate machine, I believe the code necessary for executing everything through a single script on the jenkins slave via SSH to the pod's build server would be much more convoluted and confusing. There are two separate environments that need to be setup in different ways. At a jenkins point-of-view, things are very easy,it simply needs to execute the bash script run_tests.sh and nothing else...

Patch-set: 1