Reformat auth_url based on the ID API version.
No longer raising an OSCredsError when the auth_url does
not end with a 'v' + some number. Additionally, the auth_url
will be massaged to remove any 'v' + num from the end of
the URL and generate its own version value based on the ID
API version configured.

JIRA: SNAPS-144

Change-Id: I3a7844025324105576da59b1516d0f541281e6bf
Signed-off-by: spisarski <s.pisarski@cablelabs.com>