Update patch set 1
Patch Set 1:

I do not believe this change has anything to do with these test failures. I run these tests multiple times daily and when they fail at this point, that means that floating IPs are not working properly on the pod. If you think there is something else going on here, please provide more comprehensive logs or access to a pod on which the tests are failing so I can run the tests in a debugger.

Patch-set: 1