Update patch set 1
Patch Set 1:

In general terms, a test shouldn't be disabled just because it fails. That is the purpose of testing in our CI, to show what works and what doesn't. These kind of problems can be commented in the release notes of the installer/feature project IMO.

Patch-set: 1
Label: Code-Review=0