Update patch set 1
Patch Set 1:

> In general terms, a test shouldn't be disabled just because it
 > fails. That is the purpose of testing in our CI, to show what works
 > and what doesn't. These kind of problems can be commented in the
 > release notes of the installer/feature project IMO.

The issue here is not that the test fails per se, is how it fails. It is detailed in the JIRA. The ssh comes in the VM on the first interface (via floating IP) but gets out on the second interface (because of default route not being set properly). So the test needs some added logic. Since Steven plans to change the whole approach of it, I though to just remove it for E5.0 and add it when it is properly functioning.

Patch-set: 1
Reviewer: Gerrit User 5267 <5267@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0