Update patch set 2
Patch Set 2: Code-Review+1

Ok, thanks for the clarifications. My comment was just applicable for "general terms", as we do for tempest black-list (which is empty now), but this probably makes sense to be merged since the coverage is not what the test case does as such but the Ansible coverage.

Patch-set: 2
Label: Code-Review=+1