Update patch set 7
Patch Set 7:

what about if it is a nested stack and showing all the resources in the stack and which one failed?  I want to know if the stack fails, is it because neutron port failed to be created, or nova instance or what.

Patch-set: 7
Reviewer: Gerrit User 334 <334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0