Added ext_net_name into template substitution variable.
Without the change, the template would always leverage the
external network name of 'external'.
Also am now creating a security group for the VM with the
floating IP so the test will always be able to access.
Fixed another test that used the same heat template that
was changed to add in security groups.

JIRA: SNAPS-206

Change-Id: I19b09e91a844f88bd473ea25b2bba71ae19b0aa1
Signed-off-by: spisarski <s.pisarski@cablelabs.com>