Update patch set 1
Patch Set 1:

In case of OpenStack, several projects such as rally are out of offical requirements. As you said the requiments can be overriden when gating. But in case of rally, a patch cannot be merged in case of API mismatch. Here we will see later in Functest the possible issues.
And in OPNFV, all projects follow the same rules (releases).

Patch-set: 1