Update patch set 1
Patch Set 1:

Linda, snaps-verify-master is indeed the CI job in question.

For the fine details of the CI job, see run_tests.sh in the snaps 'ci' directory. It runs with the parameter -ci which runs all of the tests that can be against XCI and uses the function add_openstack_ci_tests. Fatih really only setup the CI environment to run against the XCI pod on intel-virtual10 which is currently broken. I hope this addresses your question.

BTW, I will be merging this patch in the next hour or so. Should any of your comments on this patch be construed that you have any issues with this change?

Patch-set: 1