Added exception handling around adding security groups.
When a VM instance has a security group associated with it and
the client attempts to add the same security group, a Pike pod
now raises a ClientException where previous releases simply return
None. This patch adds the appropriate exception handling that
allows several existing tests to pass not only with Pike but also
with all OpenStack versions.

JIRA: SNAPS-239

Change-Id: I282df5a2b516dab415ba5e2e7208176f2ba3f270
Signed-off-by: spisarski <s.pisarski@cablelabs.com>