Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

The overall pattern of your enhancement looks fine. All I ask if to cleanup the overall logic in the function you altered in openstack_tests.py as my original logic was not great and has become more apparent by the addition of one more call to the NetworkConfig constructor.

One other concern here is that you have only implemented the new test parameter, network_config, to only SimpleHealthCheck. I will not reject this patch over this; however, it seems to me that you will want this feature to be implemented in all of the tests instantiating networks or at least using them. If you decide not to implement the network_config parameter in all tests, please do not close out the ticket SNAPS-252 and make notes to it accordingly.

Patch-set: 1
Label: Code-Review=-1