Update patch set 1
Patch Set 1:

Taseer & Manuel, I added you to this patch so you can see the new OpenStackVmInstance method "cloud_init_complete()" which will return T/F when checking a VM with a floating IP whether or not cloud-init has completed. I am not sure if you are issuing any commands on your VMs. If so, you will probably want to know when it is completely initialized.

Patch-set: 1