Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Can you show me where this is causing a failure? I have never encountered an issue with this test with an ID api of 2.0 and I believe it was written against a 2.0 pod.

Also, please use the keystone_utils version constant when checking for the version.

Patch-set: 1
Label: Code-Review=-1