Update patch set 1
Patch Set 1:

As this is the case, the correct fix here would be to have the keystone_utils#get_domain_by_id() function do this check and return None when the keystone client version is 'v2.0'. Then the test in question would need to be changed to ensure that None is being returned after your version check logic that you have already added. I hope this makes sense...

Patch-set: 1