Update patch set 2
Patch Set 2:

The difference is in the conditional of the elif, which has one less test than the original if. I wanted to keep my change separate from the original in case I needed to make other changes (which I didn't so far). Alternative would be to user the modified conditional of the elif as the conditional in the if.

Patch-set: 2