Update patch set 2
Patch Set 2:

Stuart, this sounds like an edge condition that would need to be handled differently. IMO, we would need to send in an additional parameter that would tell the tests to add the test's user to the admin role. I do not want to code for the lowest common denominator.

Patch-set: 2