Update patch set 2
Patch Set 2:

To initialize the snaps tests, functest sends in parameters via the method os_source_file_test#parameterize(). We currently are supporting special data structures (dict) to override the test's default values and are currently supporting flavors, and images. In this case, you could simply add in a list of roles that one would want to add to the test user. This logic can then be added to the __start__() method. I will try and put together a patch for you.

Patch-set: 2