Update patch set 3
Patch Set 3:

Dimitrios, yes you are correct. Please be aware that this structure is currently not nested like this and legacy clients (functest only) will be expecting these values to apply only to the flavor's metadata so you have a couple of choices if you want to continue with this method.

1. Make sure Functest changes it's dict accordingly
2. Support backwards compatibility through adding additional logic to the new function in openstack_tests.py

Patch-set: 3