Removed project query restriction when looking up network for port creation.
When creating a router to an external network, the port configuration object
was performing a lookup using the port's user and project ID; however, as
external networks created in other projects and are visible amongst all
projects can also be added to a port. This patch removes the project restriction
from the network query. I do not believe this fix should ultimately break other
functionality as networks that cannot be looked up by a certain user would
still return None raising a valid exception.

JIRA: APEX-598

Change-Id: I6cd32373fd4f42d8ad9b63a89c2131f95b113ce8
Signed-off-by: spisarski <s.pisarski@cablelabs.com>