Update patch set 3
Patch Set 3:

> Hi. Sorry for being late to the review...
 > @Panos: as Steve mentioned, just go ahead and push an new patch.
 > The extra_specs should not include the flavor config values as you
 > mentioned.

Hi Georg,
Yes, I've created a new test patch which is depended on this one.
I have added you as reviewer.
This one is merged and I'll continue the design as well as any needed update of Steve's patch as part of new one.
Steve also agrees about that the flavor config values should not be included in the extra_specs field and I'm following this approach. [You can see the solution of this bug as part of new test patch]

Patch-set: 3