Fixed issue when attempting to add an internal subnet to a router.
Internal subnets that are owned by the same user who is creating a
router now can be added properly. Attempted to fix an issue with the
change in https://gerrit.opnfv.org/gerrit/#/c/57853/ that was allowing
for subnets with the same name to be assigned but ended up breaking
this functionality completely.

JIRA: SNAPS-312

Change-Id: I1687f66db47520e93e401d3e9fb5f0c4f45d460f
Signed-off-by: spisarski <s.pisarski@cablelabs.com>