Update patch set 3
Patch Set 3:

> we should have backported one by one and submit them in one click.

I had the same thought at the beginning, but in the [0] document, as a good practice, it is mentioned that it could be good if we push a commit having the whole related solution in the same patch. That's why I've followed this approach.

If you disagree (maybe this is an exception of the practice [0]), I can abandon this patch and prepare 3 new ones (modifying the code where it is necessary in order to be backported only the related impact)

[0] https://wiki.openstack.org/wiki/GitCommitMessages
    (chapter : Structural split of changes)

Patch-set: 3