Update patch set 1
Patch Set 1:

Steven: I may misunderstand your logic. If the network is external or shared, a port (respectively router port or vm) can be plugged even if it's not owned by the same user. Then searching for a network in the project is not right.

The previous logic skip looking the network in the project for a shared 'and' external. One condition is enough even if both may be met.

Patch-set: 1
Label: Verified=0