Update patch set 2
Patch Set 2:

When a network can be retrieved and it is either external or shared, a port can be created against this network; however, if it is not, an additional check to ensure that the network is part of the same project is required. The way it is currently being done may not necessarily be the best but that is what is intended. If you can outline the issues you are seeing with the current logic, we can possibly find a better means to your ends.

Patch-set: 2